Moved patching to OutputChannel, out of Channel. Masking followed it.
Logging module load exceptions.

Added a Common directory to the probing path, a sibling of Modules, for non-module DLLs.  This will keep them from being attempted to be loaded and unnecessary load errors in the warning log.

Added a parameter to VixenSystem.Start to make execution startup optional to allow an administration app to get the system to start without revving up everything for execution.

I know that everything is tightly coupled to implementation and that there's no separation between data and behavior, and I'm sorry for that, but it is what it is.  In my list of desires was a decoupling of behavior and presentation, but I didn't care about data and behavior.  I know that's going to make things harder from a maintenance standpoint, but that's going to have to be a future concern.

Killed controller definitions.
(Existing controllers are now invalid)

Changed how the output module is managed for controller chains.