Added a "PropertyDependencies" member to the effect base class.  It will receive the ids of its dependencies that are properties.
When creating a CommandNode (required for the system execution entry point and created by the sequence base class), the effect's property dependencies are checked against the properties in the target nodes.  A very verbose exception is thrown if requirements are not met.

Catching exceptions thrown during repository population so that the offending module can be named in the error log instead of a general error message.

Various other little, miniscule things.