NOTICE: BREAKING CHANGE
Moved transforms from controllers down to the output module.
Why?
Because the modules currently called "file templates" (and originally intended to be so) are going to become "module templates".  They will be able to affect a module when it's created (such as adding transforms to outputs).  This will allow the templating to be centralized instead of having to try to make use of them wherever a file is created.

This is going to break any controllers you have created, and I'm sorry for that.