Changed the TimelineControl to be called a TimelineGrid -- and made it part of a "master" control called "TimelineControl". It also now has a TimelineHeader (for the times), and a TimelineRowList (for the tree on the LHS of all the rows).
It had to be broken down like this -- as best I can tell -- for the scrolling. We want the LHS to scroll vertically with the grid, but stay static during hoizontal scrolling. Vice versa for the header. Hopefully, we can get them appearing to scroll using events.

My apologies to anyone trying to read through the diff: I didn't actually change anything major -- if anything at all -- when moving TimelineControl.cs to be TimelineGrid.cs. 

Let me know if this is a horrible way to go about it; I can't see any better alternative, but then again, I'm a WinForms noob. That's the power of SVN: if it's shit, roll it back. ;-)