- added a simple cursor for the grid. If you click anywhere, it moves the curor (vertical line) there. Doesn't actually *do* anything, yet, though; it will mostly be for client use.
A few questions and requests for feedback:

should the cursor move and be updated when clicking/dragging a selection rectangle? At the moment, it only gets updated if the user clicks in a single spot.

Should it draw behind the elements? I was thinking not, as there might be many, many elements, and therefore it's impossible to see the cursor. But the way it is now it looks messy.