- added mouse wheel events, for vertical scrolling in the control. I was starting to natually use it when testing the tree, and got annoyed that it wasn't there. :-)
Although, I don't like the way it's implemented; the way the events are done and fed up through the RowList control to the main one. Is that the preferred way to do it? It doesn't look elegant, but I'm not familiar with events so can't quite tell how to do it.

Also, the way the .NET framework events work is weird; I obviously don't understand them fully. So please correct me if doing things wrong here.