- aah, that's better. Handling the mouse wheel events a *lot* nicer, now -- maybe even properly. ;-)
Turns out it *does* work if you only put the handler in the parent control, I thought it had to be in the child controls as well. The reason it wasn't working originally, though, is because the control wasn't getting selected properly, as we were not calling the base methods for all the mouse events. I've added those back in, and the controls gets selected properly when you click on it, and all is well.

It doesn't scroll if the control doesn't have focus, apparently that's by design. If we want to make it scroll when moused over, there's more info here: http://stackoverflow.com/questions/4769854/windows-forms-capturing-mousewheel