Because I'm a complete tard and love it when something works as expected...
In preparation for the context shelf set functionality we discussed on the call, each installation needs a unique identity.  GUIDs have worked well for identities so far, so I needed to get it into the system data.  Enter...file versions!

I increased the version for UserData.xml from 1 to 2 and added a migration to the reader.  And it worked!
I was so incredulously impressed that I manually reversed it and ran through the process again to make sure it really worked the first time.  My ego wanted this to be its own commit.