- changed the TimedSequenceEditor to fit with KC's recent changes. KC, does this fit with your intended model?
I'm unsure about a few things; as far as I can see, the editor module instance base doesn't really do anything. Should it still be the one to instantiate and display the form? Or is that left up to the module client/caller to do? Or, should the Start() call be static, and instantiate and display the form?