Part 3 of a series, "Conference Call Fallout".
This one implements the ability to package up the system context for use in another installation.  It packages up UserData.xml and the Modules branch.  The system-side implementation will package up the context into a file and unpackage it into a specified location, but it's up to the consumer to decide how to make this presentable to the user.

It will copy out the current context and then overlay the other context onto it so that the user of the target system can use the client application that they're familiar with (which will probably be the same across users anyway).

Any sequences the user wishes to share will have to be sent external to this process.

A possible usage example is to have the admin application maintain a collection of context files that the user has collected from other users and then provide the ability to switch between contexts.  But it would have to know the name of their client app in order to launch it for them.  It could even allow them to overwrite an old context with a new one from the same user (context object contains a "source" GUID).

The package also contains fields for a name and description for the user to provide when importing (or maybe provide it from an existing package when overwriting one).

This all makes sense in my head.

Anyway, successfully tested packaging a context, unpackaging it, and running it.  Verified that user data was brought across and intact.