As I'm trying to get stuff working, I'm breaking things by using them in ways not intended; as well as fixing it up, I figured it might be a good idea to catch those errors (if possible) to save the software throwing exceptions left, right and centre (and thus giving a bad idea about its stability). This way, we can more accurately report to developers what's going on, or what's broken in their modules; I'm sure I'm not the only one that will make silly mistakes. :-)
As such, this is a small checking with a possible suggestion for a way to work around it -- what do you think? Obviously I'll fix up my code to not create things with null parameters, but it's a good idea to see than instead of it randomly crashing. :-)  As always, if it's crap, roll it back.

- added a null check for effect parameter values when writing the sequence out to XML.