Cleaning up some of the remnants of evolution as suggested by the team.  Most of the changes are fallout from that. Notable: Command -> /dev/null CommandData -> Command CommandNode -> EffectNode
An effect is no longer a render-all singleton wrapped by a class that has the values necessary for rendering and is instead an instance with the values needed to render itself.

EffectNode has most of the changes.
It now also has a rudimentary dirty flag.
NOTE TO EDITOR IMPLEMENTORS: If you change the content of an effect's parameter or target node members, it will NOT be marked dirty and will therefore not be re-rendered with the new values.  Until we implement a more comprehensive model, please reset these references when changing values.

Verified system startup and effect rendering.