Time to get your re-org boots on.  Nah, it's not that bad.
After Jonathon and Michael's discussion and the resulting Michael and I discussion, this is a change to allow execution of overlapping effects.
Most of the touched units have minor changes (expecting an array instead of a single instance).
I've added the MultiStateEnumerator class to implement the change.  TimedChannelEnumerator is still around for reference, just not being used.

I have not verified the execution of overlapping effects mainly because we don't yet have a mechanism to create that scenario.
I have verified basic execution of serialized effects.