Hold on to your hats....
- added some color tools (a color picker, and color gradient editor) to CommonElements. Also has some cool-looking controls in ControlsEx.

I intend to use these to work with the RGB property module, and for effects to easily set up colors and colour changes. Needs to be tweaked for our use, but I wanted to get it in while it's in a pretty clean state before stuffing around with it. It builds at least, and that was enough of a hassle. :-)