VIX-13
revamp the main sequencer menu a bit. There's now "Save" and "Save As", and it prompts for a filename if needed.

There isn't a "New" or "Open" anymore, since that's all handed from the main application: I forgot about that.

(also, slight modification to the project references: it now references them as within the same solution, so it can open the code references properly. This may break it when opening in the AllModules solution, though. :-( )