VIX-13
improvements to the timed sequence editor -- mainly to support multiple instances of a channel in a single sequence. To see what I mean, set up the system with 4 channels (A, B, C, D) and a group, G, which contains the 4 channels again (A, B, C, D). So in the editor you'll see 9 rows: the channels individually, then the group row, then if you drop the group down you'll see each individual channel again. Now, drag a new effect on there -- and it should show up in both instances of the channel at the same time. Then, if you drag it around, it should behave relatively sanely.

- removed the TimelineRow ClearAllElements: it doesn't need to clear child rows, as they are referenced 'flat' from the main control.
- added checking to a timeline element to actually ensure it's moved before raising a 'moved' event.
- LINQ'd up the SelectedElements property in TimelineGrid, to make them distinct/unique. (a single element may be in multiple rows, so may be returned multiple times).
- Improved the vertical dragging in TimelineGrid to support elements in multiple rows. Keeps track of the current mouse row, which means we can do things a bit smarter.
- changed the TimedSequenceEditor to use a single element for each effect, rather than a single element for each channel an effect references. Much cleaner that way.
- fixed an XML bug when loading a sequence with multiple channel references.