VIX-13
fixed a bug when vertically dragging multiple elements, and an element: (a) didn't have any instances in the current row, (b) existed in multiple visible rows in the grid, and (c) had an instance that was dragged off the edge of the grid.

It's not ideal, but we're limited by the amount of information we have about the user's desires when dragging to different rows, and have to have a best guess at it.

Also commented some of the methods. I should comment more of them, I'll try and do that sometime.