VIX-15
added initial playback support for the editor. 3 buttons: Play, Stop, and Pause.

Also added a ProgramContext to the form, so it can pass stuff to the execution engine. KC, can you check if I'm doing that all right?

Enabled the execution when the app starts up (so we can actually *do* something!)