VIX-14
I forgot to commit before implementing the change, so this has some testbed changes mixed in with it.  Sorry about that.

This splits UserData into ModuleStore and SystemConfig.  I've tested this against the described scenario without exception and verified that the property's use and configuration were persisted.

Both files have been placed into a SystemData directory.  The root elements of the files have changed, but other than that, the content of the old files can be copied across.

I just realized that I didn't make changes to the system context stuff in response to this.  I will make a note and get that done by the end of this weekend.