VIX-17
Hadn't saved everything in VS before committing last time, whoops. So a few project file updates.

In the descriptors for the effect editors, moved the parameter list to generate it when the property is requested. When being created outside, it tries to create it in the class constructor, which crashes: since some types are dependent on other modules (ie. for curves, or the color gradient types) it can't find them, and .NET barfs.