VIX-17
Added the editors for two specific effects: the Pulse effect, and the SetLevel effect. Also added an effect editor form to the Sequencer to be able to configure effects. (all untested so far).

In the Effect Module Repository, removed the automatic mapping of module dependencies to module property dependencies, as we don't want it to be automatic: the PropertyDependencies is really more like "Property Requirements", so we want the user to be able to specify it.

Also tweaked the descriptors of some effects and editors for consistency.