VIX-15
Fixed the Effect Rendering times: it was pulling the wrong times in (using the absolute sequence start time, instead of the effect-offset start time (ie. 0). )

Also fixed the ChannelData restricting: the Restrict() uses a start time and end time, but the data up to that point is all in start time / duration.

Also added a few bits to the data logging; it'll disappear soon, but I want to have it in SVN history.