VIX-15
Yaaaaay! the execution works fully! :-)

Reimplemented a bit of the TimedChannelEnumerator, to make it optionally leading edge triggered (to send a single pulse per iterated item). This is needed because we actually need to iterate through sequences like that: when we iterate, we don't actually care how long the effect is for, we just want to push it through to the execution engine as soon as it happens (ie. leading edge triggered). So we *can* actually iterate through them, one at a time.

Changed the SequenceExecutor to use said single TimedEnumerator. (Classes will be renamed, but will do that in a separate checkin).