VIX-20
Fixed the bug with perpetual scrolling timeline controls. Turns out it was the auto-scrolling while dragging off the edge; it was having trouble with the non-zero values, which was a result of it assuming that setting VisibleTimeStart would be capped to sane values, which was removed in the refactor. So added that back in.

Also made the CursorPosition publicly available.