VIX-4
Revamped the controller config form a bit, seems a bit nicer to use now. Could still do with some improvements.

Also added a 'Outputs Editor' form, from the controllers. This can be used to rename outputs individually, or add/remove/configure transforms on each output. Still a bit buggy, but it's a start, and can be fixed up later on.

Removed the 'setup patches' form, that can be done elsewhere.