VIX-4
Fixed up the controller output config form a bit. Updates/refreshes properly now, and can also add/delete/configure transforms.

The transforms don't seem to be saving properly yet, but I think that's because it's only using the test transform so far. I'll add a new dimming one and see how that goes.

Fixed a bug in the OutputModule: it wasn't setting the output transforms properly.