VIX-3
Added support for ordering drag/drop in the MultiSelectTreeview. It works, but good luck to anyone who tries to debug it or add features in future. I seem to have a knack of _completely fucking up_ controls when I play with them. ;-)

You can now drag-drop channels around a bit, drop them into other channels to make groups, etc., and also reorder -- there's a pretty line that lets you show where it's going to go.

It doesn't actually *do* anything with channels/groups yet, since that's handled in the events that the form can catch and deal with -- but the infrastructure is there now, and it should be pretty easy.