VIX-25
We were starting each controller in a separate thread to prevent bad actors from gumming up other controllers.  However, that resulted in some odd behavior.  It didn't matter how those threads were started -- thread pool, PFX, explicit threads -- the result was the same.  There is a comment in the code about it.  Until the issue can be resolved: 

1. ControllerManager is assuming to start all controllers in the context of the UI thread.

2. Controllers must assume to never be in the UI thread (before, the fake Renard checked, but even THAT was unreliable, so now it assumes.  There is a lot of documentation on the web about InvokeRequired lying to you because it has to).

It will take someone smarter than myself to get this resolved.  All we are buying by starting in multiple threads is parallel startup -- no one controller can interfere with the startup of others.  It may be that this is not a big deal.  If someone does come up with a clean resolution, I'd love to learn what it is.