VIX-35 VIX-36
Fix the unselected-but-still displayed controller problems properly -- when updating a controller, it was reloading the list, which lost the selection, but the form was still populated with details for the controller. So that should _properly_ fix the reported problem of VIX-35.

This will fix the first few comments of VIX-36, but not the transforms crashing; will need to implement different modules to test properly.