VIX-39
Implemented the change discussed in the bug (since there was no objection).
ApplicationServices.CreateEditor handles creating the sequence object, creating the editor, creating data objects, assigning the sequence to the editor, returning the editor reference.

Michael, one of the changes involved showing the available sequences instead of the available editors.  I changed all of the applications.  If you're not cool with this change, let me know.