VIX-13
added a generic menubar to the application, to App modules to attach to (with the AppCommands property). Not ideal, but I tihnk we'll just change to a menu for now until we can revise it.

Also modified the way effects draw (with the narrower rows now) so details are more visible.