VIX-42
added curve library implementation. Booyah.

Can now 'link' a curve to one in the library, which means it will track it as long as it is linked. You can unlink it and use it as a normal curve if desired. So we can make a single curve for things (like a dimming curve), link it to each one, then they'll all be consistent -- change one, you change 'em all.

They're linked by name now (the name is the key in the library) -- so it's not possible to rename a library curve after it's made. Might regret that.