VIX-42
added gradient library implementation. Works the same as the curve, but a bit more complex underneath due to the data types worked with.

Also fixed up references for a handful of projects.

Also potentially fixed VIX-51; not sure what the problem is, but it *could* have been having a hissy fit if you loaded old sequences that didn't have the updated curve data types. Need to wait for more info from KC on what the crashes were.

there still exists a bug where effects aren't flagged as dirty if the underlying curve/gradient changes from within the library. The item itself would be updated correctly, however the effect doesn't know of that and doesn't know it needs to re-render itself. Need to look into the 'dirty' flag a bit more.