VIX-3
finished the polish up of the channel/group config form. It should mostly be done now.

Added a context menu for the treeview, to easily copy/paste things around, work with properties, etc. It could be argued that the buttons in the "operations" groupbox aren't needed anymore if it's all available through the context menu.....