VIX-49
Revamped the main application window. No more tabs, everything in a simple little form. If we need to administer stuff, we can do it via the menu bar.

Also added a current execution state for it, so we have a little status light, and an option to start/stop it from the menu strip. A few changes to the Execution class to support that.