VIX-49 VIX-55
Fixed two bugs: when reloading system config, it wouldn't load anything below the root nodes, as the static mapping of channelnode instances wasn't ever getting cleared. And, the status light on the main app form wasn't getting updated properly, as the text/icon were getting updated with states that were out-of-order, due to the Invoke() calls, cross-threading, etc.

To fix this, we make new Managers when the system is reloaded, instead of trying to reuse the old ones. (The only reason we were reusing old ones was because of the events and subscribers, so have now made that a static event in the NodeManager class). Also moved the ChannelNode instance tracking (against GUID) into the node manager class, since it's not static anymore, and is tied to the current node collection.

(some of this has come out of the work to reverse-map commands to RGB colors, which is needed for effect drawing and the display preview module. More work coming on that later.)