VIX-7 VIX-10
Added support for 'reverse mapping' channel/command data back to the source ChannelNode, and to figure out if it's RGB or not. Needed for the above two tickets: display preview integration, and also to draw elements with detail on what they're doing.

It's pretty messed up, the whole process, and is *slow as shit* in the editor: mainly the rendering of data back to a bitmap form. I'm taking the approach of "eh, screwit, burn CPU cycles for now" for the time being, though. We need to put that bitmap rendering into a separate thread, though...