I set the assembly version to an initial 3.0.0.0.  We can do something to get that modified as we release externally-available builds, just not worrying about it right now.
I also set the copyright information.  Now, I put it in my name, but let me explain why.  Being that the copyright is a legality issue and our little gang isn't a legal entity, I opted for just my name.  That also limits any exposure.
If anyone has a problem with this, please tell me.  I don't want this to seem like a glory-hound move.  I'm happy to throw your names around where it counts so that people know that you guys gave a significant contribution to all of this.