VIX-7
A few minor changes to the display preview module: got it building with out system/format/output location, changed the namespace for convention, added to main solution, renamed project (to not be "DisplayPreviewModule" but just "DisplayPreview"). Still need to rename module implementation files to fit with convention, but that's trivial.