VIX-46
Fixed the overlapping effect drawing which regressed during the undo/redo changes. We still need to send the element changed event -- even if it's currently updating -- as the Row uses it to resort the element list, and the drawing needs them in order of start time. In addition, it doesn't matter if the rows raise those events; the undo/redo will only generate actions from the events raised by the main class/control, not individual elements.

(also fixed a wrong check in the Row class).