I'm going to throw this out there and see if there are any calls for its rollback.  Not a terribly invasive change (except for a couple points), but no behavior has been changed.  Put down the pitchfork.
Explanation: I had an issue where data was being written to the system but lost along the way.  Those are a BITCH to diagnose.  It ended up being something stupid I had done with the effect, but this commit is an attempt to help us and users diagnose such problems.

What I needed was a way to determine, at a glance, where the data was in the system.  Or, more precisely, how the data being written was affecting the system.  What I came up with was a rudimentary instrumentation subsystem.  It's in its infancy, so the implementation may not be 100% correct for its purpose (though I wouldn't call it completely inelegant, but a bit so).

I created four possible types of values:

Count - It counts
Rate - Per-second rate
Percent - % qualifying of a total count
Average - Average (count per second) over the lifetime of the value

Then I created a handful of value implementations and used a few of them. The ones I used:

Total effects written to the system
Effects/second written to the system
% commands qualified by time
% commands expired by time

There are others created, but not in use yet. They are all in Vixen.System/Sys/Instrumentation.

I then created an app module that is nothing but a form that displays the four values as text, sampled every second.  The intent was to make it graphical, but this was only for proof-of-concept, so it wasn't worth the work.

You can see the invasive points:

MultiStateTimedEnumerator - Added methods so subclasses would know if an item qualified or expired.
SystemChannelEnumerator - Made use of the new protected methods to update instrumentation values it owns.
Execution - Owns and updates two instrumentation values (in _Write).
VixenSystem - Added an Instrumentation property.

The usage of the values is pretty lightweight, generally adding or incrementing a value.
The getting of a value is also generally pretty lightweight, maybe doing a couple division operations at worst.
Used as intended, I don't think it should impact performance negatively.  Please let me know if you believe differently.

Okay, pull your pitchforks back out and speak your mind.  If I've overlooked something, please call me to the mat on it.