VIX-17
added the concept of being able to tell an effect editor what effect it is editing. This is needed for a bit of extra customization and configurability; for example, a spin needs to know how long it will be operating over to calculate values. It can be used for more powerful things later on as well (such as disabling color controls if not operating on a target with color details, etc.)

Note that it is not guaranteed to be set; an effect editor can not rely on it having been set. If it's null, an effect editor should do the best it can and try to default to sane values.