VIX-17
Fixed up the spin a bit, and broke it a bit more too.

Added a 'chase' effect; sort of like a single-revolution spin, but you can also set the path the movement takes through the group (so not linear). Buggy at the moment, but I wanted to get it in before I went to bed.