<no ticket>
stripped out the 'fake renard' output module into a new one called "Dummy Lighting". Figured it's a good basic one we can keep with it for people that have no controllers and want to play with it.

You'll most likely need to do a clean and rebuild after getting this update, just to make sure.

There's now no more of the Test modules for the real one. Huzzah!