VIX-67
Pardon my French, but UNHOLY FUCK.  I have been sitting in the same position for so long, tracking down this issue, that my joints are sore.  I hope we never have to track something like this down again.  Entirely my own doing, but I think I'm allowed to swear at myself.

The transform data seems to be persisting and restoring now and other uses of module data don't seem to be negatively affected.  Let me know if this is not the case.

I'll try to outline the causes and other changes:

1. XmlSystemConfigReader.cs - In the loop where it's populating the controller's outputs, it was setting the controller's whole transform collection to each output's transform collection.  The result is that the controller ended up with the transform collection of the last output.

2. ModuleDataSet.cs - No direct cause in here, but there were a few things that were problems waiting to happen.  Some duplicated code and one place where a dataset's origin wasn't being set.

3. IOutput.cs - Again, no problems here, but there was a method that wasn't in the interface.  It's not being called by anything, but I added it in the name of a comprehensive fix.

4. OutputModuleInstanceBase.cs - SetTransforms wasn't following the correct pattern for adding transforms (calling _AddTransformsToOutput).

5. TransformModuleManagement.cs - The clone method was assuming that if the module had data, the data was not orphaned.

6. Modules.cs - More duplicate code; it was duplicating effort that is the responsibility of a dataset object and was functionality already provided by a dataset object.