Beta feedback!  I never considered someone trying to use a serial port-based controller without no serial port.  I changed the serial port config to account for that and tried to change the modules that use it (Renard and DMXUSBPro).
I ensured that the application did not blow up for either output module whether there are ports or not.  Erik, I'm sorry for diving into your code.  Feel free to call me to the mat on any of the changes I made.  But a high-profile bug like that (introduced by ME, no less) on the first night...let's say I was motivated to destroy it.