Forgive me, group, for I have sinned; this commit is the conglomeration of a number of changes.  But, on the plus side, it can be rolled back wholesale.  It implements a couple new things and changes a couple others.  I will explain and try to justify each.  You all may come up with better ideas for implementations or question the what and why and I'm open to all of it.
1. Controller chaining has been fixed.  The whole reason this existed in the first place is to allow the user to more closely model their display.  You can have Renard controllers that daisy-chain.  While they can be modeled as a single controller, chaining allows them to set it up how they actually have their hardware set up.  If we get to the point where they are doing all of this graphically, the visual representation will be more accurate this way.  I tested it using a 24-channel Renard that is actually three 8-channel controllers chained.  It worked in both configurations.

2. Moved UpdateInterval into the IOutput interface.

3. Added ChainIndex to the IOutput interface so that the module can know where it is in the chain.  This is necessary for Renards to correctly format their data.  Also added it to the OutputController class because that class is aware of the chaining whereas the output module is not.  It is only told where it should assume it is; used when the root controller of the chain has to create data for other controllers in the chain.

4. Introduced the concept of output-source collections.  The idea is that entire collection can be added or removed.  This was going to be used for something later in this commit, but that went a different direction.  It left us with a clean way to add and remove the channels as sources, so I kept it in.

5. ChannelManager no longer assumes to use SystemChannelEnumerator for the channels' enumerator.  Whatever enumerator the channels are opened with is used for any channels added after opening the channels.  The enumerators in use (only two so far) are expected to have a specific constructor signature, which is not a good, but that can be changed later.  This is part of some new end-user "testing" functionality.

6. ChannelManager was using both _instances and _channels as the definitive collection of channels.  Moved it to being just _instances.

7. ControllerManager allows adding/removing collections of sources.

8. Implemented a "test" state.  It's an end-user test state.  It's just like the normal Open state, but it opens the channels with a different enumerator -- NonExpiringChannelEnumerator.  The enumerator makes the last non-null command it pulled from a channel the everlasting state for that channel.  So if a level command is written to a channel, the channel stays at that level.

To use it, I implemented a test form (not committed) that allows the user to select an effect, nodes to apply the effect to, and specify parameters for the effect.  That gets the effect to render itself.  They can then use a trackbar to slide through the data, back and forth, sliding through the state of the effect.  This is akin to the "channel test" in 2.x that gave them sliders so they could go through the range of intensity.  But with this, as an example, they could render an RGB effect that goes through a number of colors and then slide nodes through that state.  It requires first stopping the engine then going to the test state.

The UI and code scarred up the admin app just for testing, so I didn't commit it.  But the code is available upon any request.  THEN you can tell me what a crap implementation it is.

(what number am I at?)

9. Execution._UpdateChannelStates -- I removed the condition because if there isn't anything subscribed to that effect, the channel states are never updated!

10. Removed fixing up and removing of channel sources in OutputController.  This is handled by the state transitions now.

I've run the app through as much execution as I could in both the usual execution state and the new test state and it seems to work the same as usual.  I am sorry about committing so much at once, I swore I'd try to avoid doing this again, but it was the result of circumstances.