App module that exposes some input module usage.  The name is utilitarian and it has no real use as of yet, but it's a good proof of concept.  It allows you to select an input module, select an input from that module, specify an effect for it to trigger, specify which parameter the input module will affect, and select which nodes the effect will go to.
I tried to make it _somewhat_ generic in its module and parameter handling.  I've successfully used it to make a joystick generate Renard data.  Technically cool, but pretty useless.  I've also used to it generate the new SetPosition effect against the new PSC module and make a servo respond to a joystick axis, which is a bit more useful.

The data generated by the input hardware is executed live.  I hope to get it recorded into a sequence as a next step.