Pulled pretty much the rest of the behavior of the InputEffectRouter module out into a separate class in Common.  The module just hosts it.  This leaves the behavior separate, able to be consumed by other objects without having the reference the app module.
The new class is BasicInputManagement.  You feed it the lists of input modules and input effect mappings that you have for yourself and it manages starting/stopping the modules, effect generation, and notifying you when there is new data.  It also has the forms needed for management.  Then you, as the consumer, chooses where that data goes.

This is a step towards allowing editors to make use of inputs without having to re-implement everything.  Also, the app module can now allow the user to turn on/off live execution of input-generated data.  It doesn't, but it can.

All of this allows the consumer of inputs to be very thin.  See the app module for an example.  Outside of usual app module stuff, it now only:
* Creates an instance of BasicInputManagement.InputManager.
* Invokes the management form and updates its data when the user confirms the dialog.
* Starts the input manager when it starts.
* Stops the input manager when it stops.
* Has the system execute the input-generated data when the input manager says there is data.

Hopefully, this is the final step before having an editor be able to record and playback the input-generated data.
I'm taking a break now.