VIX-104 Added consideration for command-line arguments to VixenApplication.  The two supported at this time are:
no_execution: Prevents execution from starting.
no_controllers: Disables all controllers.

If in this situation, once the application is started with the no_controllers argument, the application is in the same state as if it started normally and the controllers were manually and individually disabled.
However, if you specify no_execution, the controllers are left in the whatever enabled/disabled state is in the sys config.  So once execution is started, any controllers enabled in the sys config will be started.
If you combine the two, controllers are disabled AND execution is not started.

Also moved a try/catch block that was wrapping a _StartController call into that method since it's called from other places.