Submitted for your evaluation...
1. Bringing a "test" form to the admin app.  I put it under the "System" menu.  This may not be the best place for it.

2. ControllerManager.OpenControllers did everything in the UI thread and now CloseControllers does as well.  This allowed the dummy lighting module to not define new Hide and Show methods.  I think I'm the one that originally threaded it.  Great for performance, bad for accounting for anything that needs to be in the UI thread.

3. Added IsOpen, IsClosed, IsInTest properties to Execution.cs since it already references the state classes and other classes don't have to.  The current state string is still available, but this provides some convenience shortcuts to the currently available states.  Feel free to insult this decision.

4. Changed the "current state" indicator around a bit in the admin app so that the label could resize.  There aren't any other controls on the status bar yet, so I hope this is cool with Michael.

The idea behind the test form is that they can pick the nodes to test (so they can test a single channel or their entire display, if they wanted) and an effect to use.  An effect because A) we deal in effects and B) they may want to test by using intensity levels or color or servo positions or...you get the idea.
When doing this against multiple nodes, it does it by setting the state of each in turn.  There's a chance that you can slide the trackbar slow enough to see each channel transition separately.  I'm hoping that's not a big deal.

Like I said at the start, this is submitted for evaluation.  It's all in a single commit so it can be rolled back.