Finally have a day where I'm not at work.  God love me, it's a miracle.
Got intents working!  This code base is hugely different than the trunk.  I want to see how this all pans out before imposing my crap onto any part of the code base.  A lot of the changes in here were just to get it to compile and run.  I've been using a junk "ramp" intent and a junk "ramp" effect that uses that intent (they are accidentally in the "official" module branch; that will need to change because they're quite unofficial).

I'm going to blather for my personal posterity's sake so I can remember some of this...

For effects and intents, this introduces the idea of having the modules (through the base classes) automatically handle reading and writing parameters/values to an object array through properties in the class.  For example...the ramp effect module has StartLevel and EndLevel properties that rely on the module's data as the backing store for the values (nothing new there).  Those properties are marked with a Value attribute (better name pending).  The base class picks up those properties at instantiation.  When something writes to the effect's ParameterValues, it automatically goes to the appropriate properties.  Same for reading ParameterValues.  It relies on reflection, but it means not having every module having to manage getting values in and out of the object arrays.

There is a context created for every sequence, as always.  There is also now a "system" context that you write to for live data.

Sequences no longer write their data to the system for execution.  What happens, very coarsely, is this:
 1. Controller says "I want to update".  It calls Execution.UpdateState.
 2. Execution.UpdateState gets all contexts to update and then updates the channels' states from those contexts.  Channels are no longer containers of data.  They have a current state that is the aggregation of the contexts' states for each channel. i.e. Every context has a current state for every channel.
 3. The controller then pulls its current state from the channels.
Going from the push model to the pull model means that there is less churning during idle periods and that effects won't run-on as they do after execution.  (It currently doesn't clear the context's channel states when it's stopped, but it at least stops updating it).

Oh yeah, filtering.

No filter modules yet, but the framework is in place.  When the contexts are updated, the driver for that process separately has the context filter (so that users can select to disable all filters in the future).  This is for pre-filters.  For post-filters, a controller does something similar.  It has each output update its state from the channels and then separately has each filter its state.

Patch is dead.  Channels are now sources and outputs pull straight from them.  There is still patching, but it's being maintained by the system as a relationship between channels and outputs, separate from the channels.

My Monster drink is wearing off and I'm getting hungry.  Like I said, I'm just seeing how all of this plays out in the real world, hence the branch.  I'm only using a 64-channel test rig right now, but the ramps are smooth and responsive, which is all I hoped for at this point.